refactor(#93): 로그인 관리 코드 ViewModel로 옮김
