+++ AOS/app/src/main/java/boostcamp/and07/mindsync/ui/login/LoginEvent.kt
@@ -1,7 +1,7 @@
 package boostcamp.and07.mindsync.ui.login
 
 sealed interface LoginEvent {
-    data object Success : LoginEvent
+    data class Success(val accessToken: String) : LoginEvent
 
     data class Error(val message: String) : LoginEvent
 }
